requirements.txt
selenium==4.15.0
pandas==2.1.4
pytest==7.4.3
pyyaml==6.0.1
